Json schema support - fix test
RecordRepositoryTest#produceAndConsumeRecordUsingJsonSchema created a topic while producing a message, this influenced the other tests because number of topic was different depending on the test order. Now topic for this test is created in KafkaTestCluster to have always the same number of topic for each test regardless of the test order.