feat(helm): added jwt secret auto gen (#1062)
Introduced a new auto generated secret key that will only be
generated on first install called "jwtSignatureGeneratorSecret".
The value of this secret key is added via an environment variable
setting to the depoyment only if no user defined variable with
the same name is found. This should avoid duplicate env var names
that can cause trouble with subsequent changes to the deployment
object.