add graceful shutdown for docker
Execute java process with PID=1 through replacing the existing process with the java process. This is required for docker based execution as docker forwards SIGINT to the process with PID=1. In case the process cannot handle it, the signal is ignored and docker forcefully kills the container after a certain grace period leading to an exit code of 137.