Allow anonymous users to read default topic configuration
Previously, in AKHQ 0.24, it was possible to read the default topic
configuration when security is enabled and the user is anonymous.
Now, it is no longer possible to do so.  That is probably because
"@Secured(SecurityRule.IS_AUTHENTICATED)" was added to the
outer TopicController class as part of the security refactor in
AKHQ 0.25.

The problem being fixed here was ultimately preventing anonymous users
from creating new topics when security is enabled.  It was not possible
to even access the UI form for filling out the new topic information.

The solution is to allow anonymous users to read default topic
configuration.